# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Chat API with self-destructing messages
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  profilePicture?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  /** @nullable */
  profilePicture?: string | null;
  isOnline?: boolean;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface OnlineUsers {
  onlineUserIds: string[];
}

export interface Message {
  id: string;
  senderId: string;
  receiverId: string;
  encryptedMessage: string;
  /** @nullable */
  plainText?: string | null;
  /** @nullable */
  timer?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  isRead?: boolean;
  createdAt: string;
}

export interface MessageInput {
  receiverId: string;
  message: string;
  /**
     * Self-destruct timer in seconds (null = no timer)
     * @nullable
     */
  timer?: number | null;
}

export interface Conversation {
  user: User;
  lastMessage: Message;
  unreadCount: number;
}

export type GetUsersParams = {
/**
 * Search by username or email
 */
search?: string;
};