# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Chat API with self-destructing messages
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  senderId: string;
  receiverId: string;
  encryptedMessage: string;
  /** @nullable */
  plainText?: string | null;
  /** @nullable */
  timer?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  isRead?: boolean;
  createdAt: string;
}