# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Chat API with self-destructing messages
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  receiverId: string;
  message: string;
  /**
     * Self-destruct timer in seconds (null = no timer)
     * @nullable
     */
  timer?: number | null;
}