# lib/api-zod/src/generated/types/onlineUsers.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Chat API with self-destructing messages
 * OpenAPI spec version: 0.1.0
 */

export interface OnlineUsers {
  onlineUserIds: string[];
}