# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Chat API with self-destructing messages
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  profilePicture?: string | null;
}